# gpt/openapi.yaml
openapi: 3.1.0
info:
  title: GPT sin Alzheimer API
  version: 1.0.0
  description: API local para que GPT gestione notas, usuarios y preferencias de forma privada.

servers:
  - url: https://gptsinalzheimer.loca.lt
    description: Servidor local del usuario

paths:
  /usuarios:
    get:
      summary: Lista los usuarios disponibles
      operationId: listarUsuarios
      responses:
        '200':
          description: Lista de nombres de usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /usuario:
    post:
      summary: Crea un nuevo usuario
      operationId: crearUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
      responses:
        '200':
          description: Usuario creado

  /usuario/activar:
    post:
      summary: Activa un usuario existente
      operationId: activarUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
      responses:
        '200':
          description: Usuario activado

  /notas:
    get:
      summary: Lista las notas del usuario actual
      operationId: listarNotas
      responses:
        '200':
          description: Lista de nombres de archivo
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /nota:
    post:
      summary: Crea o actualiza una nota
      operationId: guardarNota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                contenido:
                  type: string
      responses:
        '200':
          description: Nota guardada

  /nota/{nombre}:
    get:
      summary: Devuelve el contenido de una nota
      operationId: obtenerNota
      parameters:
        - name: nombre
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contenido de la nota
          content:
            text/plain:
              schema:
                type: string

    delete:
      summary: Borra una nota por nombre
      operationId: borrarNota
      parameters:
        - name: nombre
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Nota eliminada